kinship-backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
anthropic==0.34.2
python-dotenv==1.0.1
supabase==2.4.3
